Refactor VocabularyReviewSession component: streamline state management, enhance answer handling, and integrate event dispatch for learned vocabulary tracking.
